moved logger pub-sub into server package
