+++ src/edu/saddleback/cs4b/Backend/Server/Logger.java
@@ -1,4 +1,4 @@
-package edu.saddleback.cs4b.Backend.Utilitys;
+package edu.saddleback.cs4b.Backend.Server;
 
 import edu.saddleback.cs4b.Backend.PubSub.SystemEvent;
