+++ src/edu/saddleback/cs4b/Backend/Server/ServerLogger.java
@@ -1,4 +1,4 @@
-package edu.saddleback.cs4b.Backend.Utilitys;
+package edu.saddleback.cs4b.Backend.Server;
 
 import edu.saddleback.cs4b.Backend.PubSub.Observer;
 import edu.saddleback.cs4b.Backend.PubSub.Subject;
